+++ lib/src/app_auto_launcher_impl_macos.dart
@@ -1,50 +1,46 @@
-import 'dart:io';
+import 'dart:async';
 
-import 'app_auto_launcher.dart';
+import 'package:flutter/services.dart';
+import 'package:launch_at_startup/src/app_auto_launcher.dart';
 
 class AppAutoLauncherImplMacOS extends AppAutoLauncher {
   AppAutoLauncherImplMacOS({
-    required String appName,
-    required String appPath,
-  }) : super(appName: appName, appPath: appPath);
+    required super.appName,
+    required super.appPath,
+    super.args,
+  });
 
-  File get _plistFile => File(
-      '${Platform.environment['HOME']}/Library/LaunchAgents/$appName.plist');
+  static const platform = MethodChannel('launch_at_startup');
 
   @override
   Future<bool> isEnabled() async {
-    return _plistFile.existsSync();
+    final isEnabled =
+        await platform.invokeMethod<bool>('launchAtStartupIsEnabled');
+    if (isEnabled == null) {
+      throw Exception(
+        'WARNING: AppAutoLauncherImplMacOS.isEnabled null response! platform.invokeMethod<bool>("launchAtStartupIsEnabled") returned a null response when checking if app is set to launch at startup.',
+      );
+    } else {
+      return isEnabled;
+    }
   }
 
   @override
   Future<bool> enable() async {
-    String contents = '''
-<?xml version="1.0" encoding="UTF-8"?>
-<!DOCTYPE plist PUBLIC "-//Apple//DTD PLIST 1.0//EN" "http://www.apple.com/DTDs/PropertyList-1.0.dtd">
-<plist version="1.0">
-  <dict>
-    <key>Label</key>
-    <string>$appName</string>
-    <key>ProgramArguments</key>
-    <array>
-      <string>$appPath</string>
-    </array>
-    <key>RunAtLoad</key>
-    <true/>
-  </dict>
-</plist>
-''';
-    if (!_plistFile.parent.existsSync()) {
-      _plistFile.parent.createSync(recursive: true);
+    if (!await isEnabled()) {
+      await platform
+          .invokeMethod('launchAtStartupSetEnabled', {'setEnabledValue': true});
     }
-    _plistFile.writeAsStringSync(contents);
     return true;
   }
 
   @override
   Future<bool> disable() async {
-    if (_plistFile.existsSync()) {
-      _plistFile.deleteSync();
+    if (await isEnabled()) {
+      await platform.invokeMethod(
+        'launchAtStartupSetEnabled',
+        {'setEnabledValue': false},
+      );
     }
     return true;
   }