+++ CHANGELOG.md
@@ -1,3 +1,33 @@
+## 0.3.1
+
+- fix: web build error.
+
+## 0.3.0
+
+* [windows] Add MSIX app support
+* [windows] fix: detect disabled/enabled register (#26)
+* [macos] macOS support API change to allow App Sandbox mode #27
+
+## 0.2.2
+
+* add startup args for linux (#19)
+
+## 0.2.1
+
+* feat: add interactive process type to MacOS #16
+
+## 0.2.0
+
+- fix: web compile error #14
+
+## 0.1.9
+
+- bump win32_registry to 1.0.2 #12
+
+## 0.1.8
+
+- [windows] Fix incorrect setting/checking of registry value #10
+
 ## 0.1.7
 
 - [windows] Replace `win32` with `win32_registry`